⚙️ chore(updater): 优化更新流程并引入 changes.json 控制文件 【新增】新增 changes.json 路径配置，用于判断是否需要删除旧资源 【优化】调整 cover_folder 逻辑，仅在 changes.json 不存在时删除旧资源，避免误删 【优化】统一异常提示文案，提升用户体验 【新增】清理阶段同步删除 changes.json，确保临时文件完整清理

+++ updater.py
@@ -21,6 +21,7 @@ class Updater:
 
         self.exe_path = os.path.abspath("./assets/binary/7za.exe")
         self.delete_folder_path = os.path.abspath("./assets/images")
+        self.changes_file_path = os.path.abspath("./update_temp/changes.json")
 
         self.download_file_path = os.path.join(self.temp_path, self.file_name)
         self.extract_folder_path = os.path.join(self.temp_path, self.file_name.rsplit(".", 1)[0])
@@ -43,17 +44,21 @@ class Updater:
 
     def cover_folder(self):
         """覆盖安装最新版本的文件。"""
-        print("开始覆盖安装...")
-        while True:
+        if not os.path.exists(self.changes_file_path):
             try:
                 if os.path.exists(self.delete_folder_path):
                     shutil.rmtree(self.delete_folder_path)
+            except Exception as e:
+                print(f"删除旧资源文件失败: {e}")
+        print("开始覆盖安装...")
+        while True:
+            try:
                 shutil.copytree(self.extract_folder_path, self.cover_folder_path, dirs_exist_ok=True)
                 print("覆盖安装完成")
                 break
             except Exception as e:
-                print(f"覆盖失败: {e}")
-                input("按回车键重试. . .")
+                print(f"覆盖安装失败: {e}")
+                input("按回车键重试. . . \n Press any key to continue")
 
     def terminate_processes(self):
         """终止相关进程以准备更新。"""
@@ -79,6 +84,8 @@ class Updater:
         try:
             os.remove(self.download_file_path)
             shutil.rmtree(self.extract_folder_path)
+            if os.path.exists(self.changes_file_path):
+                os.remove(self.changes_file_path)
             print("清理完成")
         except Exception as e:
             print(f"清理失败: {e}")
